+++ suppaftp/src/async_ftp/tokio_ftp/mod.rs
@@ -53,6 +53,11 @@ where
     welcome_msg: Option<String>,
     active_timeout: Duration,
     passive_stream_builder: Box<TokioPassiveStreamBuilder>,
+    /// flags whether a data connection is currently open
+    ///
+    /// Since it isn't possible to have multiple data connections at the same time,
+    /// this flag is used to track whether a data connection is currently open.
+    data_connection_open: bool,
     #[cfg(not(feature = "async-secure"))]
     marker: PhantomData<T>,
     #[cfg(feature = "async-secure")]
@@ -95,6 +100,7 @@ where
             marker: PhantomData {},
             mode: Mode::Passive,
             nat_workaround: false,
+            data_connection_open: false,
             passive_stream_builder: Self::default_passive_stream_builder(),
             welcome_msg: None,
             #[cfg(feature = "async-secure")]
@@ -152,6 +158,7 @@ where
             mode: self.mode,
             nat_workaround: self.nat_workaround,
             passive_stream_builder: self.passive_stream_builder,
+            data_connection_open: self.data_connection_open,
             tls_ctx: Some(Box::new(tls_connector)),
             domain: Some(String::from(domain)),
             welcome_msg: self.welcome_msg,
@@ -206,6 +213,7 @@ where
                     mode: Mode::Passive,
                     nat_workaround: false,
                     welcome_msg: None,
+                    data_connection_open: false,
                     passive_stream_builder: Self::default_passive_stream_builder(),
                     tls_ctx: None,
                     domain: None,
@@ -223,6 +231,7 @@ where
             reader: BufReader::new(DataStream::Ssl(stream.into())),
             mode: Mode::Passive,
             nat_workaround: false,
+            data_connection_open: false,
             passive_stream_builder: Self::default_passive_stream_builder(),
             tls_ctx: Some(Box::new(tls_connector)),
             domain: Some(String::from(domain)),
@@ -460,6 +469,7 @@ where
         debug!("Finalizing retr stream");
         // Drop stream NOTE: must be done first, otherwise server won't return any response
         drop(stream);
+        self.data_connection_open = false;
         trace!("dropped stream");
         // Then read response
         self.read_response_in(&[Status::ClosingDataConnection, Status::RequestedFileActionOk])
@@ -531,6 +541,7 @@ where
         // Drop stream NOTE: must be done first, otherwise server won't return any response
         stream.shutdown().await.map_err(FtpError::ConnectionError)?;
         drop(stream);
+        self.data_connection_open = false;
         trace!("Stream dropped");
         // Read response
         self.read_response_in(&[Status::ClosingDataConnection, Status::RequestedFileActionOk])
@@ -576,6 +587,7 @@ where
         self.perform(Command::Abor).await?;
         // Drop stream NOTE: must be done first, otherwise server won't return any response
         drop(data_stream);
+        self.data_connection_open = false;
         trace!("dropped stream");
         self.read_response_in(&[Status::ClosingDataConnection, Status::TransferAborted])
             .await?;
@@ -807,6 +819,7 @@ where
         debug!("closing data connection");
         // Drop stream NOTE: must be done first, otherwise server won't return any response
         drop(stream);
+        self.data_connection_open = false;
         trace!("dropped stream");
         // Then read response
         self.read_response_in(&[Status::ClosingDataConnection, Status::RequestedFileActionOk])
@@ -852,6 +865,10 @@ where
 
     /// Execute command which send data back in a separate stream
     async fn data_command(&mut self, cmd: Command) -> FtpResult<DataStream<T>> {
+        // guard data connection
+        self.guard_multiple_data_connections()?;
+        self.data_connection_open = true;
+
         let stream = match self.mode {
             Mode::Active => {
                 let listener = self.active().await?;
@@ -1068,6 +1085,18 @@ where
                     .map_err(FtpError::ConnectionError)
             })
         })
+    }
+
+    /// guard against multiple data connections
+    ///
+    /// If `data_connection_open` is true, returns an [`FtpError::DataConnectionAlreadyOpen`] indicating that a data connection is already open.
+    /// Otherwise, returns Ok(()).
+    fn guard_multiple_data_connections(&self) -> FtpResult<()> {
+        if self.data_connection_open {
+            Err(FtpError::DataConnectionAlreadyOpen)
+        } else {
+            Ok(())
+        }
     }
 }
 
@@ -1423,6 +1452,72 @@ mod test {
         drop(container);
     }
 
+    #[tokio::test]
+    async fn test_should_prevent_multiple_data_connections() {
+        let (mut stream, _container) = setup_stream().await;
+        let command = "LIST";
+
+        let _data_stream = stream
+            .custom_data_command(command, &[Status::AboutToSend])
+            .await
+            .expect("Failed to perform custom data command");
+        // Try to open another data connection without closing the previous one
+        match stream
+            .custom_data_command(command, &[Status::AboutToSend])
+            .await
+        {
+            Err(FtpError::DataConnectionAlreadyOpen) => {}
+            _ => panic!("Expected DataConnectionAlreadyOpen error"),
+        }
+
+        // try with all other data commands
+        match stream.retr_as_stream("somefile.txt").await {
+            Err(FtpError::DataConnectionAlreadyOpen) => {}
+            _ => panic!("Expected DataConnectionAlreadyOpen error"),
+        }
+
+        match stream.put_with_stream("somefile.txt").await {
+            Err(FtpError::DataConnectionAlreadyOpen) => {}
+            _ => panic!("Expected DataConnectionAlreadyOpen error"),
+        }
+
+        match stream.append_with_stream("somefile.txt").await {
+            Err(FtpError::DataConnectionAlreadyOpen) => {}
+            _ => panic!("Expected DataConnectionAlreadyOpen error"),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_should_free_data_connection_after_close() {
+        let (mut stream, _container) = setup_stream().await;
+        let command = "LIST";
+
+        let (response, data_stream) = stream
+            .custom_data_command(command, &[Status::AboutToSend])
+            .await
+            .expect("Failed to perform custom data command");
+        assert_eq!(response.status, Status::AboutToSend);
+        let mut reader = BufReader::new(data_stream);
+        AsyncFtpStream::get_lines_from_stream(&mut reader)
+            .await
+            .expect("Failed to get lines from stream");
+        // finalize
+        assert!(stream.close_data_connection(reader).await.is_ok());
+
+        // Now it should be possible to open another data connection
+        let (response, data_stream) = stream
+            .custom_data_command(command, &[Status::AboutToSend])
+            .await
+            .expect("Failed to perform custom data command");
+        assert_eq!(response.status, Status::AboutToSend);
+        let mut reader = BufReader::new(data_stream);
+        AsyncFtpStream::get_lines_from_stream(&mut reader)
+            .await
+            .expect("Failed to get lines from stream");
+        // finalize
+        assert!(stream.close_data_connection(reader).await.is_ok());
+    }
+
     /// Test if the stream is Send
     fn is_send<T: Send>(_send: T) {}
 
@@ -1550,10 +1645,6 @@ mod test {
     }
 
     async fn finalize_stream(mut stream: AsyncFtpStream) {
-        // Get working directory
-        let wrkdir: String = stream.pwd().await.unwrap();
-        // Remove directory
-        assert!(stream.rmdir(wrkdir.as_str()).await.is_ok());
         assert!(stream.quit().await.is_ok());
     }